Add subtitles downloading and bug fixes.

+++ main.py
@@ -3,19 +3,37 @@ import os
 import shutil
 
 import inquirer
+from inquirer.themes import load_theme_from_dict
 
-from utils import access, concat, convert, download, ext, extract, qualities, load, master, search
+from utils import access, download, ext, extract, qualities, load, master, process, search, subs
 
 MAX_WORKERS = 30  # Maximum number of workers for thread pool executor
 
 atexit.register(ext)  # Function to be executed when program exits
 
+# Theme for inquirer
+theme = load_theme_from_dict({
+    'Checkbox': {
+        'selected_color': 'cyan',
+        'selected_icon': 'x',
+        'selection_color': 'green'
+    },
+    'Question': {
+        'mark_color': 'green'
+    },
+    'List': {
+        'selection_color': 'green'
+    }
+})
+
 try:
     answer = inquirer.prompt([
         inquirer.List('type', 'Look for movies or TV shows?', ['Movies', 'TV shows'], carousel=True)
-    ], raise_keyboard_interrupt=True)['type']
+    ], theme=theme, raise_keyboard_interrupt=True)['type']
 
-    query = input('Search: ')
+    query = input('Search: \033[1m')
+
+    print('\033[0m')
 
     movie = answer == 'Movies'
 
@@ -24,7 +42,7 @@ try:
     if results:
         title = inquirer.prompt([
             inquirer.List('title', 'Select movie or TV show', results.keys(), carousel=True)
-        ], raise_keyboard_interrupt=True)['title']
+        ], theme=theme, raise_keyboard_interrupt=True)['title']
 
         data = load(results[title], movie)  # Get TV show or movie data
         acc = access(data['ID'], movie)  # Get access data
@@ -40,50 +58,58 @@ try:
 
             os.makedirs(directory, exist_ok=True)
 
-            # Try to create temporary folder, if it exists, delete it and create another
-            try:
-                os.makedirs(temporary)
-            except FileExistsError:
-                shutil.rmtree(temporary)
-                os.makedirs(temporary)
+            proceed = True
 
-            quality = inquirer.prompt([
-                inquirer.List('quality', 'Select quality', [i + 'p' for i in links.keys()], carousel=True)
-            ], raise_keyboard_interrupt=True)['quality'][:-1]
+            if os.path.isfile(os.path.join(os.getcwd(), directory, f'{title}.mp4')):
+                again = inquirer.prompt([
+                    inquirer.List('again', f'{title} has already been downloaded, download again?', ['No', 'Yes'])
+                ], theme=theme, raise_keyboard_interrupt=True)['again']
 
-            MP4 = os.path.join(directory, f'{title}.mp4')
-            TS = os.path.join(directory, f"{title}.ts")
+                if again == 'No':
+                    proceed = False
 
-            segments = extract(links[quality])  # Get segments links
+            if proceed:
+                # Try to create temporary folder, if it exists, delete it and create another
+                try:
+                    os.makedirs(temporary)
+                except FileExistsError:
+                    shutil.rmtree(temporary)
+                    os.makedirs(temporary)
 
-            download(segments, temporary, title, MAX_WORKERS)
+                quality = inquirer.prompt([
+                    inquirer.List('quality', 'Select quality', [i + 'p' for i in links.keys()], carousel=True)
+                ], theme=theme, raise_keyboard_interrupt=True)['quality'][:-1]
 
-            concat(TS)  # Join segments
-            convert(TS, MP4)  # Convert TS to MP4
+                sbs = subs(results[title], True)  # Get subtitles links
+                segments = extract(links[quality])  # Get segments links
 
-            os.unlink(TS)  # Remove TS video
+                download(segments, sbs, title, MAX_WORKERS)  # Download all segments and subtitles
+                process(directory, title, sbs)  # Join segments, convert to MP4 and add subtitles
 
-            shutil.rmtree(temporary)
+                shutil.rmtree(temporary, True)
         else:
             season = inquirer.prompt([
                 inquirer.List(
                     'season', 'Select season', [f'Season {s}' for s in data.keys() if s != 'ID'], carousel=True
                 )
-            ], raise_keyboard_interrupt=True)['season'].split()[-1]
+            ], theme=theme, raise_keyboard_interrupt=True)['season'].split()[-1]
 
             episodes = [x.split()[-1] for x in inquirer.prompt([
                 inquirer.Checkbox('episodes', 'Select episodes', [f'Episode {e}' for e in data[season].keys()])
-            ], raise_keyboard_interrupt=True)['episodes']]
+            ], theme=theme, raise_keyboard_interrupt=True)['episodes']]
 
             if episodes:
+                episodes = list(map(str, sorted(map(int, episodes))))  # Episodes might come in wrong order, so we sort
+
                 masters = {e: master(data[season][e], expiration, token, False) for e in episodes}  # Get master links
-                links = {e: qualities(link) for e, link in masters.items()}  # Get index links for each quality
+                links = {e: qualities(link) for e, link in masters.items()}  # Get links for each quality
 
                 directory = os.path.join(os.getcwd(), title, f"Season {season}")
                 temporary = os.path.join(os.getcwd(), 'temp')
 
                 os.makedirs(directory, exist_ok=True)
 
+                # Get qualities all episodes have in common
                 labels = ['360', '480', '720', '1080']
 
                 for i in labels[:]:
@@ -96,14 +122,21 @@ try:
                 if labels:
                     quality = inquirer.prompt([
                         inquirer.List('quality', 'Select quality', [i + 'p' for i in labels], carousel=True)
-                    ], raise_keyboard_interrupt=True)['quality'][:-1]
+                    ], theme=theme, raise_keyboard_interrupt=True)['quality'][:-1]
                 else:
                     quality = str()
 
                 for episode, values in links.items():
-                    # Paths for MP4 and TS videos
-                    MP4 = os.path.join(directory, f'Episode {episode}.mp4')
-                    TS = os.path.join(directory, f"Episode {episode}.ts")
+                    if os.path.isfile(os.path.join(os.getcwd(), directory, f'Episode {episode}.mp4')):
+                        again = inquirer.prompt([
+                            inquirer.List(
+                                'again', f'Episode {episode} has already been downloaded, download again?',
+                                ['No', 'Yes']
+                            )
+                        ], theme=theme, raise_keyboard_interrupt=True)['again']
+
+                        if again == 'No':
+                            continue
 
                     os.makedirs(temporary, exist_ok=True)
 
@@ -112,16 +145,13 @@ try:
                     else:
                         link = values[str(max([int(i) for i in values.keys()]))]  # Choose best quality if not available
 
+                    sbs = subs(data[season][episode], False)  # Get subtitles links
                     segments = extract(link)  # Get segments links
 
-                    download(segments, temporary, f'Episode {episode}', MAX_WORKERS)
+                    download(segments, sbs, f'Episode {episode}', MAX_WORKERS)  # Download all segments and subtitles
+                    process(directory, f'Episode {episode}', sbs)  # Join segments, convert to MP4 and add subtitles
 
-                    concat(TS)  # Join segments
-                    convert(TS, MP4)  # Convert TS to MP4
-
-                    os.unlink(TS)  # Remove TS video
-
-                    shutil.rmtree(temporary)  # Remove temporary folder and all its files
+                    shutil.rmtree(temporary, True)
             else:
                 print('No episodes selected.')
     else:
